refactor: menu domain 패키지 수정
