+++ src/main/java/kitchenpos/dao/MenuProductDao.java
@@ -1,9 +1,8 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.MenuProduct;
-
 import java.util.List;
 import java.util.Optional;
+import kitchenpos.domain.menuproduct.MenuProduct;
 
 public interface MenuProductDao {
     MenuProduct save(MenuProduct entity);
